requirements.txt
cassandra-driver==3.11.0
Flask==2.0.2
Jinja2==3.0.2
MarkupSafe==2.0.1
six==1.16.0
itsdangerous==2.0.1
Werkzeug==2.0.2